# app/static/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "News Bots V2 API",
    "description": "API for managing All routes related to News Bots",
    "version": "2.0"
  },
...
  "paths": {
...
    "/category": {
      "post": {
        "tags": [
          "Categories"
        ],
        "summary": "Create a new category for organizing bots",
        "description": "\n    this endpoint allows you to create a new category, which serves as a container for grouping related bots.\n    \n    key points:\n    - the 'name' and 'alias' fields are required. the 'name' is displayed to users, while the 'alias' is used internally.\n    - an icon url is automatically generated based on the provided alias. this icon will be used in the ui to represent the category.\n    - the 'slack_channel' field, if provided, associates the category with a specific slack channel for notifications.\n    - the 'border_color' field allows you to set a custom color (in hex format) for visual distinction in the ui.\n    - new categories are created with 'is_active' set to false by default.\n    - the 'created_at' and 'updated_at' timestamps are automatically set.\n\n    after creation, you can add bots to this category using the bot creation or update endpoints.\n    ",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "Category details",
            "required": true,
            "type": "string",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "description": "Display name of the category (required)"
                },
                "alias": {
                  "type": "string",
                  "description": "Unique identifier for the category, used for icon generation (required)"
                },
                "slack_channel": {
                  "type": "string",
                  "description": "Slack channel ID for category notifications (optional)"
                },
                "border_color": {
                  "type": "string",
                  "description": "HEX color code for category border in UI (optional, e.g., \"#FF5733\")"
                }
              },
              "required": [
                "name",
                "alias"
              ]
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Category created successfully",
            "schema": {
              "type": "object",
              "properties": {
                "success": {
                  "type": "boolean"
                },
                "data": {
                  "$ref": "#/components/schemas/Category"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input - This could occur if required fields are missing or if the alias is not unique",
            "schema": {
              "type": "object",
              "properties": {
                "success": {
                  "type": "boolean"
                },
                "error": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error - This could occur due to database issues or other server-side problems",
            "schema": {
              "type": "object",
              "properties": {
                "success": {
                  "type": "boolean"
                },
                "error": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Categories"
        ],
        "summary": "Get a category by id or name",
        "description": "This endpoint retrieves a specific category and its associated bots from the database.",
        "parameters": [
          {
            "name": "category_id",
            "in": "query",
            "description": "The ID of the category",
            "required": false,
            "type": "integer",
            "schema": {}
          },
          {
            "name": "category_name",
            "in": "query",
            "description": "The name of the category",
            "required": false,
            "type": "string",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": "Category retrieved successfully",
            "schema": {
              "type": "object",
              "properties": {
                "success": {
                  "type": "boolean"
                },
                "data": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Bad request (no valid parameters provided)",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "description": "Category not found",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/category/{category_id}": {
      "delete": {
        "tags": [
          "Categories"
        ],
        "summary": "Delete a category and its associated bots by id",
        "description": "This endpoint deletes a category and all its associated bots from the database.",
        "parameters": [
          {
            "name": "category_id",
            "in": "path",
            "description": "The ID of the category to be deleted",
            "required": true,
            "type": "integer",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": "Category deleted successfully",
            "schema": {
              "type": "object",
              "properties": {
                "success": {
                  "type": "boolean"
                },
                "data": {
                  "type": "object"
                }
              }
            }
          },
          "404": {
            "description": "Category not found",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Categories"
        ],
        "summary": "Update an existing category",
        "description": "\n    this endpoint allows you to update the details of an existing category. it's particularly useful for modifying category properties or correcting information.\n\n    key behaviors:\n    1. partial updates are supported. you only need to include the fields you want to change in the request body.\n    2. if the 'alias' is updated, the category's icon url will be automatically regenerated.\n    3. updating certain fields (like 'slack_channel' or 'run_frequency') will trigger a rescheduling of all active bots in this category.\n    4. the 'updated_at' timestamp is automatically set to the current time upon successful update.\n\n    important notes for frontend implementation:\n    - when updating the 'alias', be aware that this might change the category's icon, which could affect ui elements.\n    - if the update results in bot rescheduling, the response will include information about which bots were rescheduled or if any failed.\n    - the 'border_color' field accepts hex color codes, which can be used to update the category's visual representation in the ui.\n\n    this endpoint is crucial for maintaining accurate and up-to-date category information, which directly impacts bot organization and functionality.\n    ",
        "parameters": [
          {
            "name": "category_id",
            "in": "path",
            "description": "Unique identifier of the category to update",
            "required": true,
            "type": "integer",
            "schema": {}
          },
          {
            "name": "body",
            "in": "body",
            "description": "Updated category details",
            "required": true,
            "type": "string",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "description": "New display name for the category"
                },
                "alias": {
                  "type": "string",
                  "description": "New alias for the category (will regenerate icon URL if changed)"
                },
                "slack_channel": {
                  "type": "string",
                  "description": "New Slack channel ID for category notifications"
                },
                "border_color": {
                  "type": "string",
                  "description": "New HEX color code for category border in UI (e.g., \"#FF5733\")"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Category updated successfully",
            "schema": {
              "type": "object",
              "properties": {
                "success": {
                  "type": "boolean"
                },
                "message": {
                  "type": "string"
                },
                "data": {
                  "type": "object",
                  "properties": {
                    "category": {
                      "$ref": "#/components/schemas/Category"
                    },
                    "rescheduled_bots": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      },
                      "description": "Names of bots that were rescheduled due to the update"
                    },
                    "failed_reschedules": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      },
                      "description": "Names of bots that failed to reschedule"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Category not found - The specified category_id does not exist",
            "schema": {
              "type": "object",
              "properties": {
                "success": {
                  "type": "boolean"
                },
                "error": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error - This could occur due to database issues or problems with the scheduler",
            "schema": {
              "type": "object",
              "properties": {
                "success": {
                  "type": "boolean"
                },
                "error": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/categories": {
      "get": {
        "tags": [
          "Categories"
        ],
        "summary": "Get all available categories along with their associated bots",
        "description": "This endpoint retrieves all categories and their associated bots from the database.",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Categories retrieved successfully",
            "schema": {
              "type": "object",
              "properties": {
                "success": {
                  "type": "boolean"
                },
                "data": {
                  "type": "object",
                  "properties": {
                    "categories": {
                      "type": "array",
                      "items": {
                        "type": "object"
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "No categories found",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}